# backend/docs/swagger.yaml
basePath: /
definitions:
...
  entity.Message:
    properties:
      createdAt:
        type: string
      files:
        items:
          type: string
        type: array
      id:
        type: string
      roomId:
        type: string
      senderId:
        type: string
      text:
        type: string
    type: object
  friend.FriendResponse:
    properties:
      avatar:
        type: string
      code:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
  friend.SendFriendRequest:
    properties:
      code:
        type: string
      username:
        type: string
    type: object
  jitsi.GetTokenResponse:
    properties:
      serverUrl:
        description: URL Jitsi-сервера
        example: https://meet.jit.si
        type: string
      token:
        description: JWT для доступа к Jitsi
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: alisfrk2004@gmail.com
    name: Дарья
  description: Серверная часть веб-приложения для коммуникации.
  termsOfService: http://swagger.io/terms/
  title: Buzz API
  version: "1.0"
paths:
...
  /friends:
    get:
      description: Возвращает массив друзей текущего пользователя. Друзьями считаются
        пользователи, с которыми есть принятый запрос (status='accepted').
      produces:
      - application/json
      responses:
        "200":
          description: Список друзей (может быть пустым)
          schema:
            items:
              $ref: '#/definitions/friend.FriendResponse'
            type: array
        "401":
          description: Неавторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить список друзей
      tags:
      - friends
  /friends/{friendId}:
    delete:
      description: Удаляет связь дружбы между текущим пользователем и указанным пользователем.
        При этом также удаляется приватная комната.
      parameters:
      - description: ID пользователя, которого нужно удалить из друзей
        in: path
        name: friendId
        required: true
        type: string
      responses:
        "200":
          description: Друг удалён
        "400":
          description: Пользователи не являются друзьями
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Неавторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Пользователь не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удалить друга
      tags:
      - friends
  /friends/send-request:
    post:
      consumes:
      - application/json
      description: Создаёт новый запрос на добавление в друзья для указанного пользователя
        (по никнейму и коду).
      parameters:
      - description: Данные получателя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/friend.SendFriendRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Запрос создан (тело ответа пусто)
        "400":
          description: Некорректные данные (username/code) или попытка отправить запрос
            самому себе
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Неавторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Пользователь не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Запрос уже существует
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Отправить запрос в друзья
...
  /rooms/{id}/jitsi-token:
    get:
      description: Генерирует и возвращает JWT-токен, который позволяет подключиться
        к видеоконференции в указанной комнате. Пользователь должен быть участником
        комнаты. Токен действует ограниченное время.
      parameters:
      - description: ID комнаты
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Токен и URL сервера
          schema:
            $ref: '#/definitions/jitsi.GetTokenResponse'
        "400":
          description: Отсутствует ID комнаты
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Неавторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Пользователь не является участником комнаты
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера (например, не удалось сгенерировать
            токен)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить JWT-токен для подключения к Jitsi
      tags:
      - jitsi
...
  /rooms/{id}/text-chat:
    get:
      description: Возвращает сообщения из указанной комнаты с поддержкой пагинации
        (limit, offset). Сообщения сортируются по убыванию даты (новые сверху). По
        умолчанию limit=50, offset=0.
      parameters:
      - description: Id комнаты
        in: path
        name: id
        required: true
        type: string
      - default: 50
        description: Количество сообщений (max 100)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Смещение
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список сообщений
          schema:
            items:
              $ref: '#/definitions/entity.Message'
            type: array
        "401":
          description: Неавторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Пользователь не является участником комнаты
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить историю сообщений в комнате
      tags:
      - chat
...
  /ws/chat:
    get:
      description: Устанавливает WebSocket-соединение для обмена сообщениями в реальном
        времени. Соединение должно содержать query-параметр roomId и заголовок Authorization
        с Bearer-токеном.
      parameters:
      - description: ID комнаты
        in: query
        name: roomId
        required: true
        type: string
      responses:
        "101":
          description: WebSocket-соединение установлено
        "400":
          description: Неверный id комнаты
        "401":
          description: Неавторизован
        "403":
          description: Не является участником комнаты
      security:
      - BearerAuth: []
      summary: Подключение к WebSocket для текстового чата
      tags:
      - chat
  /ws/notifications:
    get:
      description: Устанавливает WebSocket-соединение для отправвки уведомлений в
        реальном времени. Соединение должно содержать заголовок Authorization с Bearer-токеном.
      responses:
        "101":
          description: WebSocket-соединение установлено
        "401":
          description: Неавторизован
      security:
      - BearerAuth: []
      summary: Подключение к WebSocket для уведомлений
      tags:
      - notifications
securityDefinitions:
  BearerAuth:
    description: Введите токен в формате "Bearer <token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"